Update test structure for tui
